+++ pkg/config/config.go
@@ -7,7 +7,6 @@ import (
 	"path/filepath"
 	"sort"
 	"strings"
-	"sync"
 
 	"github.com/craftcms/nitro/pkg/helpers"
 
@@ -56,7 +55,7 @@ type Config struct {
 	Sites      []Site      `json:"sites,omitempty" yaml:"sites,omitempty"`
 	File       string      `json:"-" yaml:"-"`
 
-	rw sync.RWMutex
+	// rw sync.RWMutex
 }
 
 // AllSitesWithHostnames takes the address, which is the nitro-proxy
@@ -227,7 +226,7 @@ func (s *Site) GetAbsPath(home string) (string, error) {
 }
 
 // GetContainerPath is responsible for looking at the
-// sites webroot and determing the correct path in the
+// site’s web root and determing the correct path in the
 // container. This is used for the craft and queue
 // commands to identify the location of the "craft"
 // executable.
@@ -499,9 +498,6 @@ func (c *Config) AddSite(s Site) error {
 // RemoveContainer takes a name and will remove the container by its
 // name from the config file.
 func (c *Config) RemoveContainer(container *Container) error {
-	c.rw.Lock()
-	defer c.rw.Unlock()
-
 	for k, v := range c.Containers {
 		if container.Name == v.Name {
 			c.Containers = append(c.Containers[:k], c.Containers[k+1:]...)
@@ -512,12 +508,25 @@ func (c *Config) RemoveContainer(container *Container) error {
 	return fmt.Errorf("unknown container %q", container.Name)
 }
 
+// RemoveDatabase is used to destroy or remove a database
+// engine from the config.
+func (c *Config) RemoveDatabase(database Database) error {
+	hostname, _ := database.GetHostname()
+
+	for k, v := range c.Databases {
+		h, _ := v.GetHostname()
+		if hostname == h {
+			c.Databases = append(c.Databases[:k], c.Databases[k+1:]...)
+			return nil
+		}
+	}
+
+	return fmt.Errorf("unknown database %q", hostname)
+}
+
 // RemoveSite takes a hostname and will remove the site by its
 // hostname from the config file.
 func (c *Config) RemoveSite(site *Site) error {
-	c.rw.Lock()
-	defer c.rw.Unlock()
-
 	for i, s := range c.Sites {
 		if site.Hostname == s.Hostname {
 			c.Sites = append(c.Sites[:i], c.Sites[i+1:]...)
@@ -531,9 +540,6 @@ func (c *Config) RemoveSite(site *Site) error {
 // DisableBlackfire takes a sites hostname and sets the blackfire option
 // to false. If the site cannot be found, it returns an error.
 func (c *Config) DisableBlackfire(site string) error {
-	c.rw.Lock()
-	defer c.rw.Unlock()
-
 	// find the site by the hostname
 	for i, s := range c.Sites {
 		if s.Hostname == site {
@@ -552,9 +558,6 @@ func (c *Config) DisableBlackfire(site string) error {
 // DisableXdebug takes a sites hostname and sets the xdebug option
 // to false. If the site cannot be found, it returns an error.
 func (c *Config) DisableXdebug(site string) error {
-	c.rw.Lock()
-	defer c.rw.Unlock()
-
 	// find the site by the hostname
 	for i, s := range c.Sites {
 		if s.Hostname == site {
@@ -573,9 +576,6 @@ func (c *Config) DisableXdebug(site string) error {
 // EnableBlackfire takes a sites hostname and sets the xdebug option
 // to true. If the site cannot be found, it returns an error.
 func (c *Config) EnableBlackfire(site string) error {
-	c.rw.Lock()
-	defer c.rw.Unlock()
-
 	// find the site by the hostname
 	for i, s := range c.Sites {
 		if s.Hostname == site {
@@ -593,9 +593,6 @@ func (c *Config) EnableBlackfire(site string) error {
 // EnableXdebug takes a sites hostname and sets the xdebug option
 // to true. If the site cannot be found, it returns an error.
 func (c *Config) EnableXdebug(site string) error {
-	c.rw.Lock()
-	defer c.rw.Unlock()
-
 	// find the site by the hostname
 	for i, s := range c.Sites {
 		if s.Hostname == site {
@@ -612,9 +609,6 @@ func (c *Config) EnableXdebug(site string) error {
 
 // Save takes a file path and marshals the config into a file.
 func (c *Config) Save() error {
-	c.rw.Lock()
-	defer c.rw.Unlock()
-
 	// make sure the file exists
 	if _, err := os.Stat(c.File); os.IsNotExist(err) {
 		dir, _ := filepath.Split(c.File)
@@ -629,7 +623,6 @@ func (c *Config) Save() error {
 	if err != nil {
 		return err
 	}
-	defer f.Close()
 
 	// unmarshal
 	data, err := yaml.Marshal(&c)
@@ -642,7 +635,7 @@ func (c *Config) Save() error {
 		return err
 	}
 
-	return nil
+	return f.Close()
 }
 
 func (c *Config) createFile(dir string) error {