+++ pkg/prompt/prompt.go
@@ -249,18 +249,18 @@ func CreateSite(home, dir string, output terminal.Outputer) (*config.Site, error
 		found = "web"
 	}
 
-	// set the webroot
+	// set the web root
 	site.Webroot = found
 
-	// prompt for the webroot
-	root, err := output.Ask("Enter the webroot for the site", site.Webroot, ":", nil)
+	// prompt for the web root
+	root, err := output.Ask("Enter the web root for the site", site.Webroot, ":", nil)
 	if err != nil {
 		return nil, err
 	}
 
 	site.Webroot = root
 
-	output.Success("using webroot", site.Webroot)
+	output.Success("using web root", site.Webroot)
 
 	// prompt for the php version
 	versions := phpversions.Versions
@@ -313,7 +313,13 @@ func RunApply(cmd *cobra.Command, args []string, force bool, output terminal.Out
 	for _, c := range cmd.Root().Commands() {
 		// set the apply command
 		if c.Use == "apply" {
-			return c.RunE(c, args)
+			// run the apply command
+			if err := c.RunE(c, args); err != nil {
+				return err
+			}
+
+			// call the post run command to cleanup
+			return c.PostRunE(c, args)
 		}
 	}