+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Release Notes for Craft Nitro
 
+## Unreleased
+
+### Added
+- Added the `db destroy` command, for removing database engines. ([#325](https://github.com/craftcms/nitro/issues/325))
+- Added the ability to extend the Nginx configuration with a `nitro.conf` file at the root of a project. ([#18](https://github.com/craftcms/nitro/issues/18))
+
+### Changed
+- Added support for validating multiple hostnames with the `alias` command. ([#341](https://github.com/craftcms/nitro/issues/341))
+- The `stop` command no longer returns an error code when no containers are running. ([#336](https://github.com/craftcms/nitro/issues/336))
+- The `add` and `create` commands now set `DB_USER` and `DB_PASSWORD` variables in the `.env` file. ([#310](https://github.com/craftcms/nitro/issues/310))
+- The `ls` command now outputs the internal and external ports for containers.
+
+### Deprecated
+- Deprecated the `clean` command. `apply` will automatically remove unused containers.
+- Deprecated the `validate` command.
+
+### Fixed
+- Fixed a bug where the `apply` command would not remove unused containers unless it was run twice. ([#321](https://github.com/craftcms/nitro/issues/321))
+- Fixed a bug where the `container new` command didn’t allow entering a port. ([#318](https://github.com/craftcms/nitro/issues/318))
+
 ## 2.0.7 - 2021-04-06
 
 ### Added